fixed hw1
